requirements.txt
allure-pytest==2.9.42
allure-python-commons==2.9.42
atomicwrites==1.4.0
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
ddt==1.4.2
et-xmlfile==1.1.0
idna==2.10
importlib-metadata==4.0.1
iniconfig==1.1.1
jsonpath==0.82
openpyxl==3.0.7
packaging==20.9
pluggy==0.13.1
py==1.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.4
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
zipp==3.4.1
